renamed science camera interface
